Prepare for building a task-enabled buildah image
The buildah image did not work for the use in the Tekton task. When
comparing the currently referenced image with the produced one, there
are differences around the user, entrypoint, and workspace. After this
PR, I will onboard a new component to produce a buildah image for use in
tasks. There will be a nudging relationship between these images so that
we can utilize the same binaries in both images.

Signed-off-by: arewm <arewm@users.noreply.github.com>