Simplify building of buildah
The two-step process of building buildah is over complicating matters.

We could create a container that just runs the buildah binary or we can
just create a container with the binary and run buildah from it.

This latter approach looks like what we did with the previous buildah
image which did not have its source tracked and it also is what is done
in the registry.access.redhat.com/ubi9/buildah:latest image.

Since this approach is what I was doing with the second step, let's just
remove the first step and copy the newer binary directly into a buildah
image as that is the part that we want to update.

Signed-off-by: arewm <arewm@users.noreply.github.com>