Remove entitlement sharing from host
Results for being able to install packages were inconsistent when the
entitlements were shared from the host to the container build. Since
the containers running on unentitled hosts, the sharing resulted in
errors to find appropriate packages. Users should be able to easily
install RPMs that are in UBI repos.

Resolves: KFLUXBUGS-1506

Signed-off-by: arewm <arewm@users.noreply.github.com>