Update the task image for use on remote and local builds
As part of build-definitions#1216, we need to have a version of buildah
that can work for both local and remote builds. This ports over the
changes needed for the remote runner while still being able to use the
latest buildah images.

The prior containerfile was https://github.com/konflux-ci/build-tasks-dockerfiles/blob/main/multi-platform-runner/Dockerfile

Signed-off-by: arewm <arewm@users.noreply.github.com>