Update all images to Fedora:40
New buildah commits have moved to golang:1.22. We need to update to a
new version of Fedora in order to get the newest version. Since we are
already using fedora-minimal as the final parent, we will reuse that for
the builder.

While quay.io/centos/centos:stream9 was considered as an alternative,
multiple required RPMs were not available to install.

Signed-off-by: arewm <arewm@users.noreply.github.com>