Parameterize the icm injection script inputs and outputs
The second parameter is the location of the prefetch SBOM

Since we need to inject the content sets into the Containerfile with an
additional COPY command, we don't want to require that the sbom is in
the current directory in case that pollutes the container image's build
context. Instead, we parameterize it as the location may vary between
applications of this script.

The third parameter is the output directory

Since we need to be able to copy content-sets.json into the image, we
have to ensure that the file is present within the build's context. The
output directory should be set to that location.

Signed-off-by: arewm <arewm@users.noreply.github.com>