feat: reduce unnecessary push builds
We do not need to rebuild the buildah images if the only changes are to
the Tekton pipelines. These changes will automatically be picked up once
we have substantive changes that are pushed to the images.

By not rebuilding on these changes, we can more easily enable an
auto-release flow that does not result in the potential follow-on
updates to the buildah-related tasks.

Signed-off-by: arewm <arewm@users.noreply.github.com>