Migrate buildah task image to UBI10 (#163)
* Add kubectl as a git submodule

The image will get migrated to UBI10, which does not contain
`kubernetes-client` in its repositories.
Add the Kubernetes repository
as a git submodule to build kubectl from source

Signed-off-by: Tomáš Nevrlka <tnevrlka@redhat.com>

* Build kubectl from source in buildah-task image

Drop the kubectl RPM because it's not included in UBI10 repositories.

Rename dockerfile-json-builder to a more generic dep-builder and build
kubectl in it

Run the buildah-task PipelineRuns on a change to kubectl directory too

Signed-off-by: Tomáš Nevrlka <tnevrlka@redhat.com>

* Migrate buildah task image to UBI10

Containerfile.buildah and Containerfile.image_build are already migrated
to UBI10.
Containerfile.task references a buildah image built by Konflux - use one
that is already migrated to UBI10

Signed-off-by: Tomáš Nevrlka <tnevrlka@redhat.com>

---------

Signed-off-by: Tomáš Nevrlka <tnevrlka@redhat.com>